Commented out the line in the output report for player LED
This is a fork of version 2.2.2 where I just commented out the lines concerning the player LED. Have not tested full functionality but the LED is turned off and I have not changed anything else so strange behaviour different to the original program is not expected. Do let me know how it goes on for you, if there are any technical issues I won't be able to solve them, I don't understand the program completely I just experimented.